Api Additions
vectors.worldToCameraPos(vec) takes a vector in world space and returns the vector in view space of the viewer's camera.
renderer.isCameraBackwards() checks if you're in the backwards-facing F5 view.
renderer.getCameraPos() returns a vector in world space of the location of the viewer's camera.
Modified some methods in entity api:
getPos(), getRot(), and getBodyYaw() can now take delta as an argument, giving you the lerped version between ticks.